feat: 权限解耦

+++ handlers/news_delete.go
@@ -3,6 +3,7 @@ package handlers
 import (
 	"financial_management/consts"
 	"financial_management/service"
+	"financial_management/setting"
 	"financial_management/util"
 	"github.com/gin-gonic/gin"
 )
@@ -16,17 +17,16 @@ func NewsDelete(c *gin.Context) {
 		util.Response(c, consts.ParamErrorCode, nil)
 		return
 	}
+	if userID != setting.Config.Admin.ID {
+		util.Response(c, consts.PermissionErrorCode, nil)
+		return
+	}
 	newsList := service.GetNewsByIDs(newsIDs)
 	if newsList == nil {
 		util.Response(c, consts.ParamErrorCode, nil)
 		return
 	}
-	for _, news := range newsList {
-		if news.UserID != userID {
-			util.Response(c, consts.PermissionErrorCode, nil)
-			return
-		}
-	}
+
 	if err := service.DeleteNews(newsIDs); err != nil {
 		util.Response(c, consts.SystemErrorCode, nil)
 		return
